feat(@xen-orchestra/rest-api): expose dashboard endpoint
